fix: lost error handling and notify meta update

+++ internal/rootcoord/ddl_callbacks_truncate_collection.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/milvus-io/milvus-proto/go-api/v2/milvuspb"
 	"github.com/milvus-io/milvus/internal/distributed/streaming"
+	"github.com/milvus-io/milvus/pkg/v2/log"
 	"github.com/milvus-io/milvus/pkg/v2/proto/messagespb"
 	"github.com/milvus-io/milvus/pkg/v2/streaming/util/message"
 	"github.com/milvus-io/milvus/pkg/v2/util/funcutil"
@@ -77,18 +78,28 @@ func (c *DDLCallback) truncateCollectionV2AckCallback(ctx context.Context, resul
 	}
 
 	// Drop segments that were updated before the flush timestamp
-	err := c.mixCoord.DropSegmentsByTime(ctx, header.CollectionId, flushTsList)
-	if err != nil {
+	if err := c.mixCoord.DropSegmentsByTime(ctx, header.CollectionId, flushTsList); err != nil {
 		return errors.Wrap(err, "when dropping segments by time")
 	}
 
 	// manually update current target to sync QueryCoord's view
-	err = c.mixCoord.ManualUpdateCurrentTarget(ctx, header.CollectionId)
-	if err != nil {
+	if err := c.mixCoord.ManualUpdateCurrentTarget(ctx, header.CollectionId); err != nil {
 		return errors.Wrap(err, "when manually updating current target")
 	}
 
-	return c.meta.TruncateCollection(ctx, result)
+	if err := c.meta.TruncateCollection(ctx, result); err != nil {
+		if errors.Is(err, errAlterCollectionNotFound) {
+			log.Ctx(ctx).Warn("truncate a non-existent collection, ignore it", log.FieldMessage(result.Message))
+			return nil
+		}
+		return errors.Wrap(err, "when truncating collection")
+	}
+
+	// notify datacoord to update their meta cache
+	if err := c.broker.BroadcastAlteredCollection(ctx, header.CollectionId); err != nil {
+		return errors.Wrap(err, "when broadcasting altered collection")
+	}
+	return nil
 }
 
 // truncateCollectionV2AckOnceCallback is called when the truncate collection message is acknowledged once
@@ -102,6 +113,10 @@ func (c *DDLCallback) truncateCollectionV2AckOnceCallback(ctx context.Context, r
 	// we need to forbid the compaction of current collection here.
 	collectionID := msg.Header().CollectionId
 	if err := c.meta.BeginTruncateCollection(ctx, collectionID); err != nil {
+		if errors.Is(err, errAlterCollectionNotFound) {
+			log.Ctx(ctx).Warn("begin to truncate a non-existent collection, ignore it", log.FieldMessage(result.Message))
+			return nil
+		}
 		return errors.Wrap(err, "when beginning truncate collection")
 	}
